// src/app/interfaces/formacion/planestudios/planestudiosresumen.ts
export interface Planestudiosresumen {
    creditoCasillaUno: string;
    creditoCasillaDos: string;
    creditoCasillaTres: string;
    creditoCasillaCuatro: string;
    creditoCasillaCinco: string;
    creditoCasillaSeis: string;
    creditoCasillaSiete: string;
    creditoCasillaOcho: string;
    creditoCasillaNueve: string;
    creditoCasillaDiez: string;
    creditoCasillaTotal: string;
    porcentajeCasillaUno: string;
    porcentajeCasillaDos: string;
    porcentajeCasillaTres: string;
    porcentajeCasillaTotal: string;
    key$?: string;
}